requirements.txt
asgiref==3.5.2
boto3==1.24.7
botocore==1.27.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
decorator==5.1.1
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dynaconf==3.1.9
hvac==0.11.2
idna==3.3
importlib-metadata==4.11.4
Jinja2==3.1.2
jmespath==1.0.0
loguru==0.6.0
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
s3transfer==0.6.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
validators==0.20.0
win32-setctime==1.1.0
zipp==3.8.0
